Update gradle plugin names to be clearer
