+++ common/monitoring/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("module-plugin")
+    id("node-module-plugin")
 }
 
 kotlin {
