+++ gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/BinaryPlugin.kt
@@ -4,10 +4,10 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 
-class BackendBinaryPlugin : Plugin<Project> {
+class NodeBinaryPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         target.run {
-            plugins.apply("module-plugin")
+            plugins.apply("node-module-plugin")
             extensions.configure(KotlinMultiplatformExtension::class.java) {
                 js(IR) {
                     binaries.library()
@@ -39,6 +39,28 @@ class BackendBinaryPlugin : Plugin<Project> {
                         from(file(rootProject.layout.projectDirectory.file("credentials-gcp-app.json")))
                         into("${binaryRootDirectory()}/productionExecutable")
                     }
+                    copy {
+                        from(file(layout.projectDirectory.file("Dockerfile")))
+                        into(binaryRootDirectory())
+                    }
+                }
+            }
+        }
+    }
+}
+
+class BrowserBinaryPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.run {
+            plugins.apply("base-browser-plugin")
+            extensions.configure(KotlinMultiplatformExtension::class.java) {
+                js(IR) {
+                    binaries.executable()
+                }
+            }
+            // Copy the output binaries to their final destination
+            tasks.named("assemble") {
+                doLast {
                     copy {
                         from(file(layout.projectDirectory.file("Dockerfile")))
                         into(binaryRootDirectory())
