--- gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/FrontendBinaryPlugin.kt
@@ -1,27 +1,0 @@
-package com.gchristov.thecodinglove.gradleplugins
-
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-
-class FrontendBinaryPlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        target.run {
-            plugins.apply("base-browser-plugin")
-            extensions.configure(KotlinMultiplatformExtension::class.java) {
-                js(IR) {
-                    binaries.executable()
-                }
-            }
-            // Copy the output binaries to their final destination
-            tasks.named("assemble") {
-                doLast {
-                    copy {
-                        from(file(layout.projectDirectory.file("Dockerfile")))
-                        into(binaryRootDirectory())
-                    }
-                }
-            }
-        }
-    }
-}
